# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedLedger API",
    "description": "End-to-end encrypted vault, sharing, and grant API.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/pow/verify": {
      "post": {
        "tags": ["auth"],
        "summary": "Pow Verify",
        "description": "Verify a PoW solution. Returns success or 400 on invalid solution.",
        "operationId": "pow_verify_api_auth_pow_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/POWVerifyRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register",
        "description": "Register a new user. Sends a verification code to the provided email.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify Email",
        "description": "Verify the email address with the 6-digit code sent during registration.",
        "operationId": "verify_email_api_auth_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyEmailRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "tags": ["auth"],
        "summary": "Resend Verification",
        "description": "Resend the email verification code.",
        "operationId": "resend_verification_api_auth_resend_verification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "description": "Authenticate with email + password.\nReturns tokens on success, or requires_totp=True if TOTP is enabled.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/verify-totp-login": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify Totp Login",
        "description": "Complete login with TOTP second factor.",
        "operationId": "verify_totp_login_api_auth_verify_totp_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyTOTPLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["auth"],
        "summary": "Refresh",
        "description": "Rotate a refresh token — returns a new access + refresh token pair.",
        "operationId": "refresh_api_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Logout",
        "description": "Logout from the current device.",
        "operationId": "logout_api_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogoutRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Change Password",
        "description": "Change password — requires current password. Revokes all sessions.",
        "operationId": "change_password_api_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/request-password-reset": {
      "post": {
        "tags": ["auth"],
        "summary": "Request Password Reset",
        "description": "Send a password reset code to an email address.",
        "operationId": "request_password_reset_api_auth_request_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/confirm-password-reset": {
      "post": {
        "tags": ["auth"],
        "summary": "Confirm Password Reset",
        "description": "Verify reset code and set a new password.",
        "operationId": "confirm_password_reset_api_auth_confirm_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/totp/confirm": {
      "post": {
        "tags": ["auth"],
        "summary": "Totp Confirm",
        "description": "Confirm TOTP setup with a live 6-digit code from the authenticator app.",
        "operationId": "totp_confirm_api_auth_totp_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmTOTPRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/auth/totp/disable": {
      "post": {
        "tags": ["auth"],
        "summary": "Totp Disable",
        "description": "Disable TOTP — requires password + current TOTP code.",
        "operationId": "totp_disable_api_auth_totp_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DisableTOTPRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "password": { "type": "string", "title": "Password" },
          "totp_code": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Totp Code"
          }
        },
        "type": "object",
        "required": ["email", "password"],
        "title": "LoginRequest"
      },
...
      "MessageResponse": {
        "properties": {
          "message": { "type": "string", "title": "Message" },
          "detail": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "MessageResponse"
      },
...
      "POWVerifyRequest": {
        "properties": {
          "challenge_id": { "type": "string", "title": "Challenge Id" },
          "solution": { "type": "string", "title": "Solution" }
        },
        "type": "object",
        "required": ["challenge_id", "solution"],
        "title": "POWVerifyRequest"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refresh_token": { "type": "string", "title": "Refresh Token" }
        },
        "type": "object",
        "required": ["refresh_token"],
        "title": "RefreshTokenRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "username": { "type": "string", "title": "Username" },
          "password": { "type": "string", "title": "Password" },
          "full_name": { "type": "string", "title": "Full Name" },
          "signing_public_key": {
            "type": "string",
            "title": "Signing Public Key"
          },
          "exchange_public_key": {
            "type": "string",
            "title": "Exchange Public Key"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "password",
          "full_name",
          "signing_public_key",
          "exchange_public_key"
        ],
        "title": "RegisterRequest"
      },
      "RegisterResponse": {
        "properties": {
          "user": { "$ref": "#/components/schemas/UserResponse" },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Registration complete. A verification code has been sent to your email."
          }
        },
        "type": "object",
        "required": ["user"],
        "title": "RegisterResponse"
      },
...
      "ResendVerificationRequest": {
        "properties": {
          "user_id_hex": { "type": "string", "title": "User Id Hex" }
        },
        "type": "object",
        "required": ["user_id_hex"],
        "title": "ResendVerificationRequest"
      },
...
      "VerifyEmailRequest": {
        "properties": {
          "user_id_hex": { "type": "string", "title": "User Id Hex" },
          "code": { "type": "string", "title": "Code" }
        },
        "type": "object",
        "required": ["user_id_hex", "code"],
        "title": "VerifyEmailRequest"
      },
...
    }
  }
}